+++ src/app/pages/curriculum/curriculum.component.ts
@@ -1,15 +1,19 @@
-import { ChangeDetectionStrategy, Component, inject } from '@angular/core';
-import { rxResource, toSignal } from '@angular/core/rxjs-interop';
+import {
+  ChangeDetectionStrategy,
+  Component,
+  computed,
+  effect,
+  inject,
+  signal,
+} from '@angular/core';
 
-import { map, tap } from 'rxjs/operators';
-
-import { ActivatedRoute } from '@angular/router';
 import { AsideComponent } from '../../shared/aside/aside.component';
 import { CustomTranslateService } from '../../services/custom-translate.service';
 import { FooterComponent } from '../../shared/footer/footer.component';
 import { HeaderComponent } from '../../shared/header/header.component';
 import { MainComponent } from '../../shared/main/main.component';
 import { IdentificationService } from '../../services/identification.service';
+import { IdentificationResponse } from '../../interfaces';
 
 @Component({
   imports: [AsideComponent, FooterComponent, HeaderComponent, MainComponent],
@@ -18,20 +22,17 @@ import { IdentificationService } from '../../services/identification.service';
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class CurriculumComponent {
-  private readonly activatedRoute = inject(ActivatedRoute);
   private readonly customTranslateService = inject(CustomTranslateService);
   private readonly identificationService = inject(IdentificationService);
 
-  private readonly activatedRoute$ = this.activatedRoute.paramMap.pipe(
-    map((paramMap) => paramMap.get('lang') ?? ''),
-    tap((lang) => this.customTranslateService.chageLanguage(lang))
-  );
-  private readonly currentLang = toSignal(this.activatedRoute$, {
-    initialValue: this.customTranslateService.currentLang(),
-  });
+  identificationData = signal<IdentificationResponse | undefined>(undefined);
+  currentLang = computed(() => this.customTranslateService.currentLang());
 
-  identificationResource = rxResource({
-    request: () => ({}),
-    loader: ({ request }) => this.identificationService.getIdentification(),
+  changeLangEffect = effect(() => {
+    this.identificationService
+      .getIdentification(this.currentLang())
+      .subscribe((resp) => {
+        this.identificationData.set(resp);
+      });
   });
 }