+++ src/app/shared/aside/aside.component.ts
@@ -3,7 +3,6 @@ import {
   Component,
   inject,
   input,
-  signal,
 } from '@angular/core';
 import { DomSanitizer } from '@angular/platform-browser';
 
@@ -11,10 +10,11 @@ import { TranslatePipe } from '@ngx-translate/core';
 
 import { IdentificationResponse } from '../../interfaces';
 import { NetworksComponent } from './networks/networks.component';
+import { MatIconModule } from '@angular/material/icon';
 
 @Component({
   selector: 'app-aside',
-  imports: [TranslatePipe, NetworksComponent],
+  imports: [TranslatePipe, NetworksComponent, MatIconModule],
   templateUrl: './aside.component.html',
   styleUrl: './aside.component.scss',
   changeDetection: ChangeDetectionStrategy.OnPush,