+++ src/app/shared/header/header.component.spec.ts
@@ -1,8 +1,20 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { signal } from '@angular/core';
+import { MatIconModule } from '@angular/material/icon';
+import { Router } from '@angular/router';
+
+import {
+  provideTranslateService,
+  TranslateLoader,
+  TranslatePipe,
+} from '@ngx-translate/core';
 
 import { HeaderComponent } from './header.component';
 import { identificationResponseMock } from '../../mocks/data/identification-response.mock';
-import { signal } from '@angular/core';
+import { CustomTranslateService } from '../../services/custom-translate.service';
+import { CustomTranslateMockService } from '../../mocks/services/custom-translate-mock.service';
+import { HttpClient, provideHttpClient, withFetch } from '@angular/common/http';
+import { httpLoaderFactory } from '../../app.config';
 
 describe('HeaderComponent', () => {
   let component: HeaderComponent;
@@ -10,7 +22,22 @@ describe('HeaderComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [HeaderComponent],
+      imports: [MatIconModule, TranslatePipe],
+      providers: [
+        Router,
+        provideHttpClient(withFetch()),
+        provideTranslateService({
+          loader: {
+            provide: TranslateLoader,
+            useFactory: httpLoaderFactory,
+            deps: [HttpClient],
+          },
+        }),
+        {
+          provide: CustomTranslateService,
+          useClass: CustomTranslateMockService,
+        },
+      ],
     }).compileComponents();
 
     fixture = TestBed.createComponent(HeaderComponent);
@@ -27,16 +54,19 @@ describe('HeaderComponent', () => {
     expect(component).toBeTruthy();
   });
 
-  it('should have a header with name and surname an a role in HTML', () => {
+  it('should have a header with name and surname and a role in HTML', () => {
     const htmlHeader = fixture.nativeElement.querySelector('header');
     expect(htmlHeader).toBeTruthy();
 
+    expect(component.nameAndSurname).toBeTruthy();
+    expect(component.role).toBeTruthy();
+
     const htmlNameAndSurname = fixture.nativeElement.querySelector(
-      '.header__name-and-surname'
+      '.header-data__name-and-surname'
     );
     expect(htmlNameAndSurname).toBeTruthy();
 
-    const htmlRole = fixture.nativeElement.querySelector('.header__role');
+    const htmlRole = fixture.nativeElement.querySelector('.header-data__role');
     expect(htmlRole).toBeTruthy();
   });
 });