+++ src/app/shared/main/components/summary/summary.component.ts
@@ -3,10 +3,11 @@ import {
   Component,
   computed,
   inject,
+  input,
 } from '@angular/core';
 import { rxResource } from '@angular/core/rxjs-interop';
-import { MatIconModule } from '@angular/material/icon';
 
+import { MatIconModule } from '@angular/material/icon';
 import { TranslatePipe } from '@ngx-translate/core';
 
 import { MainService } from '../../../../services/main.service';
@@ -19,11 +20,13 @@ import { MainService } from '../../../../services/main.service';
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class SummaryComponent {
+  lang = input.required<string>();
+
   mainService = inject(MainService);
 
   summaryResource = rxResource({
-    request: () => ({}),
-    loader: ({ request }) => this.mainService.getSummary(),
+    request: () => ({ lang: this.lang() }),
+    loader: ({ request }) => this.mainService.getSummary(request.lang),
   });
 
   summary = computed(() => this.summaryResource.value()?.summary);