+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 Changelog project
 
+## [1.1.0] - 2025-05-05
+
+- Added icons (angular, phone and mail).
+- Added language selector to change language dinamically.
+- Added animations loading pages.
+- Change SEO texts.
+
 ## [1.0.2] - 2025-05-02
 
 - Added robots.txt.